Split player A states into separate files
