+++ src/game-engine/application-states/PlayerA/ReadyToDeploy.ts
@@ -1,0 +1,8 @@
+import BasePlayerA from './BaseState';
+
+export default class ReadyToDeploy extends BasePlayerA {
+  constructor({ channel, stake, balances, deploymentTransaction }) {
+    super(channel, stake, balances);
+    this.transaction = deploymentTransaction;
+  }
+}
