+++ src/game-engine/application-states/PlayerA/ReadyToSendPostFundSetupA.ts
@@ -1,0 +1,13 @@
+import BasePlayerA from './BaseState';
+
+export default class ReadyToSendPostFundSetupA extends BasePlayerA {
+  constructor({ channel, stake, balances, adjudicator, signedPostFundSetupAMessage }) {
+    super(channel, stake, balances);
+    this.adjudicator = adjudicator;
+    this.message = signedPostFundSetupAMessage;
+  }
+
+  get shouldSendMessage() {
+    return true;
+  }
+}
