+++ src/game-engine/application-states/PlayerA/WaitForBlockchainDeploy.ts
@@ -1,0 +1,7 @@
+import BasePlayerA from './BaseState';
+
+export default class WaitForBlockchainDeploy extends BasePlayerA {
+  constructor({ channel, stake, balances }) {
+    super(channel, stake, balances);
+  }
+}
