+++ src/game-engine/application-states/PlayerA/WaitForResting.ts
@@ -1,0 +1,22 @@
+import BasePlayerA from './BaseState';
+
+export default class WaitForResting extends BasePlayerA {
+  constructor({
+    channel,
+    stake,
+    balances,
+    adjudicator,
+    aPlay,
+    bPlay,
+    result,
+    salt,
+    signedRevealMessage,
+  }) {
+    super(channel, stake, balances, signedRevealMessage);
+    this.aPlay = aPlay;
+    this.bPlay = bPlay;
+    this.result = result; // win/lose/draw
+    this.salt = salt;
+    this.adjudicator = adjudicator;
+  }
+}
